Honor LogAppendTime when decoding messages
When LogAppendTime is in use, set MaxTimestamp as the timestamp for all
Records in the RecordBatch. FirstTimestamp + TsDelta always form the
CreateTime, which is unwanted in this case.

There's no explicit mention in the Kafka documentation. See the official
Java client implementation:

https://github.com/apache/kafka/blob/1f1179ea64bbaf068d759aae988bd2a6fe966161/clients/src/main/java/org/apache/kafka/common/record/DefaultRecordBatch.java#L558
https://github.com/apache/kafka/blob/1f1179ea64bbaf068d759aae988bd2a6fe966161/clients/src/main/java/org/apache/kafka/common/record/DefaultRecord.java#L330-L331

The fetch test now uses additional topic with the LogAppendTime
configuration enabled. The test checks that the create timestamp set by
the producer is different from the timestamp in the fetched message. The
create timestamp is lowered by 1 millisecond to guarantee that it's in
the past and that Kafka's log append time is greater.